Cree un gráfico de barras apiladas dinámico y flexible

Esta función proporciona excelentes herramientas y opciones para la visualización de un gráfico de barras apiladas con el objetivo de mostrar el tamaño relativo (como porcentaje) de una variable de categorías, subdivididas por colores en función de un subgrupo. Dicha gráfica se va a representar usando la librería Highcharter, la cual usa internamente JavaScript.

Plot.Apiladas(
  datos,
  categoria,
  anos,
  periodo,
  colores,
  titulo = "",
  addPeriodo = TRUE,
  estilo = NULL
)

Arguments

datos

Un data frame, no un vector numérico.

categoria

Igual uso que en Plot.Series()

anos

Igual uso que en Plot.Torta()

periodo

Igual uso que en Plot.Torta()

colores

Igual uso que en Plot.Series()

titulo

Igual uso que en Plot.Series()

addPeriodo

Igual uso que en Plot.Torta()

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados para graficar las barras apiladas y cuyo objetivo es personalizar pequeños detalles de éste.

LegendTitle: Cadena de caracteres indicado un título para la leyenda (diferentes niveles del argumento varPrincipal).

hc.Tema y hc.Credits: Igual uso que en Plot.Series()

Value

Retorna el diagrama de barras apiladas (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y adicionalmente pertenecerá a la clase "highchart".

Examples

Msj <- "Se considera \u00fanicamente los valores obtenidos en el primer periodo acad\u00e9mico de cada a\u00f1o."
Plot.Apiladas(
  datos     = ejConsolidadoGrad,
  categoria = "NIVEL",      # Pruebe también con -> unique(ejConsolidadoGrad$Variable)
  anos      = c(2018:2020),
  periodo   = 1,
  colores   = c("#FFA700", "#C10AA1", "#01CDFE", "#00FF44", "#FF0040"),
  titulo    = "BARRAS APILADAS EN FUNCI\u00d3N DEL NIVEL ACAD\u00c9MICO Y EL A\u00d1O",
  estilo    = list(LegendTitle = "NIVEL ACAD\u00c9MICO:", hc.Tema = 4, hc.Credits = Msj)
)