Cree un gráfico profundo (drill down) de torta/barras dinámico y flexible

Esta función proporciona excelentes herramientas y opciones para la visualización de un gráfico drill down con el objetivo de poder inspeccionar los datos con mayor nivel de detalle, sin la necesidad de navegar o salir de él, pudiendo hacer clic en diversos elementos como columnas o sectores circulares. Dicha gráfica se va a representar usando la librería Highcharter, la cual usa internamente JavaScript.

Uso

Plot.Drilldown(
  datos,
  varPrincipal,
  varSecundaria,
  ano,
  periodo,
  torta = TRUE,
  vertical = TRUE,
  colores,
  colores2,
  titulo = "",
  label = "",
  textInfo = "",
  addPeriodo = TRUE,
  estilo = NULL
)

Argumentos

datos

Un data frame, no un vector numérico.

varPrincipal

Una variable categórica dentro del data frame ingresado en datos.

varSecundaria

Otra variable categórica dentro del data frame ingresado en datos, diferente a la principal, pues se segregará a otros niveles.

ano

Igual uso que en Plot.Torta()

periodo

Igual uso que en Plot.Torta()

torta

Si es TRUE (valor predeterminado) el primer nivel o gráfico principal será un diagrama de torta, defínalo en FALSE si desea que éste sea un gráfico de barras.

vertical

Si es TRUE (valor predeterminado) indicará que tanto la orientación del gráfico principal como secundario será vertical. Solamente aplicará si el argumento torta es FALSE.

colores

Cadena de caracteres indicando los colores con los cuales se deben colorear cada una de las trazas correspondiente a cada nivel del argumento varPrincipal. Si no se introduce algún vector se usará la paleta rainbow por defecto.

colores2

Igual que colores pero aplicado al gráfico secundario.

titulo

Igual uso que en Plot.Series()

label

Cadena de caracteres indicando el agregado al que hace referencia el gráfico. Por defecto no se emplea ningún rótulo.

textInfo

Cadena de caracteres indicando el texto que aparecerá dentro de la caja de información al pasar el mouse por las diferentes columnas del gráfico de barras.

addPeriodo

Igual uso que en Plot.Torta()

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados para graficar el drill down y cuyo objetivo es personalizar pequeños detalles de éste.

LegendTitle: Cadena de caracteres indicado un título para la leyenda (diferentes niveles del argumento varPrincipal).

hc.Tema y hc.Credits: Igual uso que en Plot.Series()

Valor

Retorna el diagrama drill down (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y adicionalmente pertenecerá a la clase "highchart".

Ejemplos

# library(dplyr)
df <- ejMiniConsolidadoAsp |>
  filter(Clase != "Sin Información", tolower(Clase) != "no aplica")
text <- "DISTRIBUCI\u00d3N DE ASPIRANTES A PREGRADO EN SITUACI\u00d3N DE DISCAPACIDAD"
Msj  <- paste(
  "Discapacidad: Deficiencia, limitaci\u00f3n de la actividad ",
  "y la restricci\u00f3n de la participaci\u00f3n."
)
Plot.Drilldown(
  datos         = df,
  varPrincipal  = "DISCAPACIDAD",
  varSecundaria = "TIPO_DISC",
  ano           = max(df$YEAR),
  periodo       = slice(df, n())$SEMESTRE,
  torta         = TRUE, # Pruebe poniendo ambos valores ahora en FALSE
  vertical      = TRUE,
  colores       = c("#FF0040", "#00FF40"),
  colores2      = c("#66C2A5", "#FC8D62", "#8DA0CB", "#E78AC3", "#A6D854", "#FFD92F"),
  titulo        = text,
  label         = "Aspirantes",
  textInfo      = "Aspirantes con discapacidades por tipo",
  addPeriodo    = TRUE,
  estilo        = list(hc.Tema = 7, hc.Credits = Msj)
)