Cree un gráfico de radar dinámico y flexible con dos diferentes paquetes

Esta función proporciona excelentes herramientas y opciones para la visualización de un gráfico de radar (también conocido como gráfico de araña) dinámico con el objetivo de observar datos multivariados de forma bidimensional. Dicho radar chart o spider plot se puede representar usando dos diferentes librerías que son Plotly y ECharts, las cuales usan internamente JavaScript.

Plot.Radar(
  datos,
  categoria,
  variables,
  estadistico = c("Promedio", "Mediana", "Varianza", "SD", "CV", "Min", "Max"),
  colores,
  rango,
  ordinal = FALSE,
  titulo = "",
  libreria = c("plotly", "echarts"),
  estilo = NULL
)

Arguments

datos

Un data frame, se espera en formato de microdatos no un agregado.

categoria

Igual uso que en Plot.Series()

variables

Una lista (ya sea creada con la sintaxis base o tidy) con las variables numéricas (mínimo tres para que se pueda realizar el gráfico) dentro del data frame ingresado en datos.

estadistico

Cadena de caracteres que indica el estadístico a graficar. Los valores permitidos son "Promedio" (valor predeterminado), "Mediana", "Varianza", "SD", "CV", "Min" y "Max".

colores

Cadena de caracteres indicando los colores con los cuales se deben colorear cada una de las trazas correspondiente a cada nivel del argumento categoria. Si no se introduce algún vector se usará la paleta rainbow por defecto.

rango

Vector numérico de longitud dos que indica el valor mínimo y máximo, respectivamente. Si no conoce el dominio del estadístico seleccionado puede usar el truco de pasar c(Min, NaN) como rango.

ordinal

Igual uso que en Plot.Barras()

titulo

Igual uso que en Plot.Series()

libreria

Cadena de caracteres que indica el paquete con el cual se realizará el radar. Los valores permitidos son "plotly" (valor predeterminado) o "echarts". Los valores se emparejarán parcialmente.

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados de acuerdo con la librería especificada para graficar el radar y cuyo objetivo es personalizar pequeños detalles de éste.

ply.LegendTitle, ply.LegendPosition y ply.Credits: Igual uso que en Plot.Series()

ply.Relleno: Cadena de caracteres indicando cómo se debe rellenar el área, toself (valor predeterminado) conecta los puntos de la traza de forma cerrada y superpone las áreas, mientras que tonext deja visible la capa más profunda y si se comparten áreas no las superpone; finalmente especifique none si desea ver únicamente los polígonos y que no se rellene el área dentro de ellos.

ply.Opacidad: Un número entre \([0, 1]\) que indica la opacidad de los polígonos/trazos.

e.Credits: Cadena de caracteres indicando el subtítulo del gráfico principal. Para mayor información consulte la función e_title().

e.Forma: Cadena de caracteres indicando el tipo de renderizado del radar, los valores admitidos son polygon (valor predeterminado) y circle. Para mayor información consulte la función e_radar_opts().

e.Tema: Modifica el tema con el cual se creará el gráfico. Los posibles valores son un número entero entre \([1, 14]\) el cual hace referencia a diferentes temas disponibles en dicha librería (helianthus, azul, inspired, macarons, westeros, walden, roma, royal, fruit, dark, chalk, purple-passion, vintage y essos respectivamente). El tema por defecto se logra al no ingresar valor alguno. Para más información consulte aquí.

e.LegType: Cadena de caracteres indicando el tipo de leyenda, los valores admitidos son plain (valor predeterminado) y scroll (útil cuando es necesario mostrar demasiados elementos). Para mayor información consulte la función e_legend().

e.LegLoc: Valor numérico o cadena de caracteres indicando la distancia entre la leyenda y el lado derecho del contenedor, puede ser expresado como un valor puntual o un valor porcentual relativo al ancho del contenedor.

Value

Retorna el radar (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente a la clase "plotly" o "echarts4r".

Examples

if (require("dplyr")) {
  Plot.Radar(
    datos     = ejSaberPro2020,
    categoria = TIPO_COL,
    variables = vars(PUNT_LECT_CRIT, PUNT_RAZO_CUANT, PUNT_INGLES),
    colores   = c("#2ACE82", "#FE2667", "#32E7C8", "#FF8D00"),
    rango     = c(0, NaN),
    estilo    = list(ply.Relleno = "tonext")
  )
}
#> Warning: ¡Se usará como estadístico la media muestral ('mean') por defecto!
#> Warning: ¡Se usará la librería 'plotly' por defecto para realizar el plot!
#> Joining, by = "TIPO_COL"

Plot.Radar(
  datos     = ejSaberPro2020,
  categoria = SEDE_NOMBRE_ADM,
  variables = vars(
    PUNTAJE_GLOBAL, PUNT_RAZO_CUANT, PUNT_INGLES,
    PUNT_LECT_CRIT, PUNT_COMP_CIUD, PUNT_COMU_ESCR
  ),
  rango     = c(0, NaN)
)
#> Warning: ¡Se usará como estadístico la media muestral ('mean') por defecto!
#> Warning: ¡Se usará la librería 'plotly' por defecto para realizar el plot!
#> Joining, by = "SEDE_NOMBRE_ADM"

Plot.Radar(
  datos     = ejSaberPro2020,
  categoria = SEDE_NOMBRE_ADM,
  variables = vars(
    PUNTAJE_GLOBAL, PUNT_RAZO_CUANT, PUNT_INGLES,
    PUNT_LECT_CRIT, PUNT_COMP_CIUD, PUNT_COMU_ESCR
  ),
  rango     = c(0, NaN),
  libreria  = "echarts"
)
#> Warning: ¡Se usará como estadístico la media muestral ('mean') por defecto!
#> Joining, by = "SEDE_NOMBRE_ADM"

misColores <- c(
  "#29ABE2", # AZUL CLARO  | Amazonia
  "#8CC63F", # VERDE       | Bogota
  "#CC241D", # ROJO        | Caribe
  "#0071BC", # AZUL VIVO   | Manizales
  "#F15A24", # NARANJA     | Medellin
  "#FBB03B", # AMARILLO    | Orinoquia
  "#93278F", # MORADO      | Palmira
  "#8A381A" # GRIS        | Tumaco
)
Msj <- "Gr\u00e1fico de radar para representar los puntajes multivariados de la prueba Saber Pro."

Plot.Radar(
  datos       = ejSaberPro2020,
  categoria   = SEDE_NOMBRE_ADM,
  variables   = vars(
    PUNTAJE_GLOBAL, PUNT_RAZO_CUANT, PUNT_INGLES,
    PUNT_LECT_CRIT, PUNT_COMP_CIUD, PUNT_COMU_ESCR
  ),
  estadistico = "SD",
  colores     = misColores,
  rango       = c(0, NaN),
  titulo      = "SPIDER PLOT",
  libreria    = "plotly",
  estilo      = list(
    ply.LegendTitle = "SEDE:", ply.LegendPosition = list(x = 0, y = -0.15, orientation = "h"),
    ply.Relleno = "tonext", ply.Opacidad = 0.8, ply.Credits = list(x = 0.7, y = -0.1, text = Msj)
  )
)
#> Joining, by = "SEDE_NOMBRE_ADM"

Plot.Radar(
  datos       = ejSaberPro2020,
  categoria   = SEDE_NOMBRE_ADM,
  variables   = vars(
    PUNTAJE_GLOBAL, PUNT_RAZO_CUANT, PUNT_INGLES,
    PUNT_LECT_CRIT, PUNT_COMP_CIUD, PUNT_COMU_ESCR
  ),
  estadistico = "CV",
  colores     = misColores,
  rango       = c(0, 0.25),
  titulo      = "RADAR CHART",
  libreria    = "echarts",
  estilo      = list(
    e.Credits = Msj, e.Forma = "circle", e.Tema = 10,
    e.LegType = "scroll", e.LegLoc = 0
  )
)
#> Joining, by = "SEDE_NOMBRE_ADM"