Cree un gráfico circular/torta/pie dinámico y flexible con dos diferentes paquetes

Esta función permite mostrar de forma interactiva una descripción compacta y general de una variable con sus respectivas categorías. Dicho diagrama se puede representar usando dos diferentes librerías que son Highcharter y Plotly, las cuales usan internamente JavaScript.

Plot.Torta(
  datos,
  categoria,
  ano,
  periodo,
  colores,
  titulo = "",
  addPeriodo = TRUE,
  libreria = c("highcharter", "plotly"),
  estilo = NULL
)

Arguments

datos
Un data frame.

categoria
Una variable categórica dentro del data frame ingresado en datos.

ano
Valor numérico que indica el año de interés para realizar el gráfico. Si solo se introduce el año, pero no el período, se calculará independiente del semestre (tomando ambos).

periodo
Valor numérico que indica el período o semestre de interés para realizar el gráfico. Si solo se introduce el período, pero no el año, se calculará en base a todos los años disponibles en datos.

colores
Igual uso que en Plot.Series()

titulo
Igual uso que en Plot.Series()

addPeriodo
Si es TRUE (valor predeterminado ssi se introduce los argumentos ano y periodo simultáneamente) concatenará al título introducido una cadena de caracteres indicando entre paréntesis el "año-periodo" insertado.

libreria
Igual uso que en Plot.Series()

estilo
Lista compuesta por varios parámetros, los cuales van a ser usados de acuerdo con la librería específica para graficar la torta y cuyo objetivo es personalizar pequeños detalles de ésta.

LegendTitle, hc.Tema, hc.Credits y ply.Credits: Igual uso que en Plot.Series()

ply.Leyenda: Por defecto la gráfica muestra la legenda fuera del grafico de pie, si se introduce la cadena de texto "inside" se resumirá toda la información dentro del pie.

Value

Retorna el diagrama circular (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente a la clase "highchart" o "plotly".

Details

Al usar el paquete Highcharter y usar las opciones de descarga, el nombre del archivo descargado será la concatenación del plot graficado y la categoría usada, así, por ejemplo, si se graficó el diagrama de pie para la categoría "Sexo" el nombre será PlotTorta_Sexo.png.

Note

Los gráficos circulares son una forma muy mala de mostrar información. El ojo es bueno para juzgar medidas lineales y malo para juzgar áreas relativas. Un gráfico de barras o un gráfico de puntos es una forma preferible de mostrar este tipo de datos.

Examples

ano <- 2020; semestre <- 2
col <- c("#F15A24", "#8CC63F")
Msj <- "Distribuci\u00f3n de estudiantes admitidos en el segundo periodo acad\u00e9mico del 2020."
Plot.Torta(datos      = Consolidado,
           categoria  = "TIPO_NIVEL",
           ano        = ano,
           periodo    = semestre,
           colores    = col,
           titulo     = "DISTRIBUCI\u00d3N DE GRADUADOS POR MODALIDAD DE FORMACI\u00d3N",
           addPeriodo = TRUE,
           libreria   = "highcharter",
           estilo     = list(LegendTitle = "\u00c9sta es una descripci\u00f3n para la leyenda:",
                             hc.Tema = 4, hc.Credits = Msj)
           )
Plot.Torta(datos      = Consolidado,
           categoria  = "TIPO_NIVEL",
           ano        = ano,
           periodo    = semestre,
           colores    = col,
           titulo     = "DISTRIBUCI\u00d3N DE GRADUADOS POR MODALIDAD DE FORMACI\u00d3N",
           addPeriodo = FALSE,
           libreria   = "plotly",
           estilo     = list(ply.Leyenda = "inside",
                             ply.Credits = list(x = 0.62, y = 1.1,
                                                text = paste0("<b>", Msj, "</b>")))
           )