Cree un diagrama de árbol (treemap) dinámico y flexible con diversos paquetes

Esta función proporciona excelentes herramientas y opciones para la visualización de datos jerárquicos/estructurados como un conjunto de rectángulos anidados. Cada grupo está representado por un rectángulo, cuya área (tamaño) es proporcional a su frecuencia absoluta (recuento) y el color se usa para mostrar otra dimensión numérica. Usando la interactividad, es posible representar varias dimensiones/niveles: grupos, subgrupos, etc. Dicha gráfica se va a representar usando la librería Highcharter, Plotly, d3treeR, entre otras, las cuales usa internamente JavaScript.

Plot.Treemap(
  datos,
  variables,
  atributo,
  textFreq = "N",
  metodo = c("Classic", "Classic2", "Sunburst", "Sunburst2"),
  estadistico = c("Promedio", "Mediana", "Varianza", "SD", "CV", "Min", "Max"),
  colores,
  titulo = "",
  libreria = c("highcharter", "plotly"),
  estilo = NULL
)

Arguments

datos

Un data frame, no un vector numérico.

variables

Una lista (ya sea creada con la sintaxis base o tidy) con las variables categóricas dentro del data frame ingresado en datos con las que se desea crear la jerarquía (recuerde que esta se crea de izquierda a derecha, es decir la primera hace referencia al grupo, la segunda al subgrupo, etc.).

atributo

Una variable numérica dentro del data frame ingresado en datos. Este es opcional y solo aplica en el caso de un nivel, es decir, cuando se ingresa únicamente una variable.

textFreq

Cadena de caracteres indicando el nombre que se le va a dar al recuento en cada uno de los grupos. Por defecto se emplea el rótulo "N".

metodo

Cadena de caracteres indicando el diseño con el cual se realizará el gráfico (en el caso de ingresar dos niveles o más). Los valores permitidos son "Classic" (valor predeterminado), "Classic2", "Sunburst" y "Sunburst2", así se usará las funciones d3tree(), d3tree2(), sunburst() y sund2b() respectivamente.

estadistico

Igual uso que en Plot.Mapa()

colores

Igual uso que en Plot.Series(), con algunos matices, cuando usamos Highcharter y nos encontramos en el caso de un nivel y especificamos un atributo es recomendable pasarle una escala de colores, pues con esto se construirá la barra horizontal. En el caso de usar el argumento atributo se puede ingresar el nombre de una paleta, por ejemplo "Set1".

titulo

Igual uso que en Plot.Series()

libreria

Igual uso que en Plot.Torta(), con algunos matices, pues en el caso de ingresar más de una variable categórica se omitirá dicho argumento, ya que metodo tomará su lugar.

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados para graficar el treemap y cuyo objetivo es personalizar pequeños detalles de éste.

hc.Tema y hc.Credits: Igual uso que en Plot.Series()

hc.borderRadius: Un número entero positivo que indica el radio del borde de cada elemento. El valor por defecto es 0 (rectángulos).

ply.Opacidad: Igual uso que en Plot.Radar()

ply.Credits: Igual uso que en Plot.Series()

sun.Explanation: Cadena de caracteres indicando qué es lo que se desea ver en el centro del sunburst al pasar el mouse por los diferentes anillos (niveles de la jerarquía). Los valores permitidos son "All" (valor predeterminado), "Count" y "Percent". Solo aplica para cuando método = Sunburst.

sun.Color: A diferencia del argumento colores acá puede pasar una paleta o vector de colores sin que se recorte (más no se recicle) éste a la longitud de categorías del nodo padre. Su uso reemplaza el funcionamiento del argumento colores. Solo aplica para cuando método = Sunburst.

sun.showLabels: Si es FALSE (valor predeterminado) no se mostrará etiquetas en los cortes. Solo aplica para cuando método = Sunburst2.

sun.colorRoot: Cadena de caracteres que indica el color del nodo raíz (root). Puede indicar el color con el nombre ("red"), código hexadecimal ("#FF0000") o RGB (rgb(1, 0, 0)). El valor por defecto es "rojo". Solo aplica para cuando método = Sunburst2.

Value

Retorna el treemap (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente a la clase "highchart", "plotly", " d3tree", "d3tree2", "sunburst" o "sund2b".

Details

Si está trabajando en un R Markdown o un aplicativo Shiny no se puede usar de forma conjunta el método = Classic (ó Classic2) y método = Sunburst (ó Sunburst2), pues se trata de un problema interno, ya que usan versiones diferentes de d3, puede darle seguimiento al problema aquí. De igual forma, si utiliza la librería sunburstR en algunas ocasiones se le verán afectadas las tablas creadas con DT.

Examples

Msj <- "Acompa\u00f1ado del Estad\u00edstico seleccionado para la Variable Edad."
if (require("viridis")) {
  Plot.Treemap(
    datos       = ejGraduados,
    variables   = SEDE_NOMBRE_MAT,
    atributo    = EDAD_MOD,
    textFreq    = "Tamaño de la Muestra",
    estadistico = "Max",
    colores     = inferno(10),
    titulo      = "TOTAL DE GRADUADOS POR SEDE DE LA UNIVERSIDAD NACIONAL",
    libreria    = "highcharter",
    estilo      = list(hc.Tema = 7, hc.borderRadius = 20, hc.Credits = Msj)
  )
}
#> Loading required package: viridis
#> Loading required package: viridisLite
#> 
#> Attaching package: 'viridisLite'
#> The following objects are masked from 'package:pals':
#> 
#>     cividis, inferno, magma, plasma, turbo, viridis
#> 
#> Attaching package: 'viridis'
#> The following objects are masked from 'package:pals':
#> 
#>     cividis, inferno, magma, plasma, turbo, viridis

Plot.Treemap(
  datos       = ejGraduados,
  variables   = FACULTAD,
  atributo    = EDAD_MOD,
  textFreq    = "n",
  estadistico = "CV",
  colores     = turbo(10, direction = -1),
  titulo      = "TOTAL DE GRADUADOS POR FACULTAD EN LA UNAL",
  libreria    = "plotly",
  estilo      = list(ply.Credits = list(x = 0.6, y = 1, text = Msj))
)

if (FALSE) {
misColores <- c(
  "#29ABE2", # AZUL CLARO  | Amazonia
  "#8CC63F", # VERDE       | Bogota
  "#CC241D", # ROJO        | Caribe
  "#0071BC", # AZUL VIVO   | Manizales
  "#F15A24", # NARANJA     | Medellin
  "#FBB03B", # AMARILLO    | Orinoquia
  "#93278F", # MORADO      | Palmira
  "#8A381A"  # GRIS        | Tumaco
)
if (require("dplyr")) {
  Plot.Treemap(
    datos     = ejGraduados,
    variables = vars(SEDE_NOMBRE_MAT, FACULTAD, PROGRAMA),
    metodo    = "Classic",
    colores   = misColores # "Set3"
  )
  Plot.Treemap(
    datos     = ejGraduados,
    variables = vars(SEDE_NOMBRE_MAT, FACULTAD, PROGRAMA),
    metodo    = "Classic2",
    colores   = "Set2"
  )

  Plot.Treemap(
    datos     = ejGraduados,
    variables = vars(SEDE_NOMBRE_MAT, FACULTAD, PROGRAMA),
    metodo    = "Sunburst",
    colores   = misColores,
    estilo    = list(sun.Explanation = "All")
  )
  Plot.Treemap(
    datos     = ejGraduados,
    variables = vars(SEDE_NOMBRE_MAT, FACULTAD, PROGRAMA),
    metodo    = "Sunburst",
    # colores   = misColores,
    estilo    = list(sun.Explanation = "All", sun.Color = list(range = RColorBrewer::brewer.pal(11, "Spectral")))
  )

  Plot.Treemap(
    datos     = ejGraduados,
    variables = vars(SEDE_NOMBRE_MAT, FACULTAD, PROGRAMA),
    metodo    = "Sunburst2"
  )
  Plot.Treemap(
    datos     = ejGraduados,
    variables = vars(SEDE_NOMBRE_MAT, FACULTAD, PROGRAMA),
    metodo    = "Sunburst2",
    colores   = misColores,
    estilo    = list(sun.showLabels = TRUE, sun.colorRoot = "#EF0055")
  )
}
}