Cree fácilmente un widget para visualización de tablas HTML usando el paquete DT

Esta función está diseñada para facilitar la creación de tablas para informes y publicaciones produciendo un widget HTML para visualizar un data frame utilizando el paquete DT. La forma en que esta función maneja las cosas por usted significa que a menudo no tiene que preocuparse por los pequeños detalles para obtener un resultado impresionante y listo para usar.

Tabla(
  datos,
  categoria,
  encabezado = "Encabezados de los Niveles de la Categoría",
  leyenda = NULL,
  tituloPdf = NULL,
  mensajePdf = "",
  ajustarNiveles = TRUE,
  colorHead = "#FFFFFF",
  colorear = FALSE,
  estilo
)

Arguments

datos
Un data frame.

categoria
Una variable categórica dentro del data frame ingresado en datos.

encabezado
Cadena de caracteres que describe los distintos niveles de la variable categoria.

leyenda
Cadena de caracteres que describe información adicional de la tabla, ésta se sitúa en la parte inferior de la tabla de manera centrada, dicho texto se visualizará en todas las opciones de descarga. Su valor por defecto es NULL.

tituloPdf
Cadena de caracteres que proporciona un título a la tabla al momento de generar el .pdf como al darle al botón de imprimir. Su valor por defecto es el introducido en el argumento encabezado.

mensajePdf
Cadena de caracteres que proporciona un mensaje situado entre el título y la tabla. Se visualizará tanto al generar el .pdf como al darle al botón de imprimir.

ajustarNiveles
Si es TRUE (valor predeterminado) se buscará optimizar el espacio entre las columnas, colocando todos los nombres de las columnas de forma horizontal y eliminado al máximo el espacio entre éstas.

colorHead
Cadena de caracteres que indica el color de fondo de la cabecera de la tabla. Puede indicar el color con el nombre ("red"), código hexadecimal ("#FF0000") o RGB (rgb(1, 0, 0)). El valor por defecto es "blanco" ("#FFFFFF").

colorear
Si es TRUE indica que desea poner color a la tabla de forma automática, aplicando color a la fuente del periodo y añadiendo un color de fondo para los años (de acuerdo con una paleta predefinida). El valor por defecto es FALSE.

estilo
Una lista compuesta por dos parámetros:

PaletaYear: Vector de caracteres que específica los colores de fondo para los años.

PaletaSemestre: Vector de caracteres que específica los colores de fuente para los periodos (semestres).

Value

Retorna la tabla creada mediante DT la cual pertenece a la clase "datatables" y "htmlwidget".

Details

Esta función se basa enteramente del paquete DT, el cual proporciona una interfaz para R a la biblioteca DataTables de JavaScript. Los data frames de R se pueden mostrar como tablas en páginas HTML, proporcionando opciones de filtrado, paginación, clasificación y muchas otras características en las tablas.

Examples

Colorido <- c("#F9CA00", "#AEF133", "#19C89B", "#FF7F50", "#D99588", "#FF4C94",
              "#19EE9F", "#9F61F6", "#F24467", "#F68118", "#F9CA00", "#AEF133")
Tabla(datos = Consolidado, categoria = "SEDE_NOMBRE_ADM",
      encabezado = "TOTAL DE ESTUDIANTES POR SEDE DE GRADUACI\u00d3N",
      titulo  = "ESTUDIANTES GRADUADOS POR SEDE",
      leyenda = "Distribuci\u00f3n de estudiantes admitidos (desde el 2009-I al 2020-II) por sede.",
      colorHead = "#8CC63F",
      estilo    = list(PaletaYear = Colorido, PaletaSemestre = c("#7E1BC3", "#C31576"))
      )

#> Joining, by = c("YEAR", "SEMESTRE")