Cree el consolidado a partir de los microdatos con los que dispone, agrupando por las variables temporales que considere necesario

Esta función permite transformar desde la base de microdatos para obtener el consolidado final con el cual trabaja la mayoría de las funciones gráficas disponibles en UnalR.

Agregar(
  formula,
  frecuencia,
  datos,
  intervalo,
  textNA = "Sin Información",
  ask = TRUE
)

Arguments

formula

Fórmula en la que el primer componente especificado (antes del \(\sim\)) hace referencia a la variable de interés, y el segundo componente (luego del \(\sim\)) a la(s) variable(s) temporales por la cuales se quiere agrupar (separadas por cualquier operador como +, *, :, etc.).

frecuencia

Vector o lista (dependiendo de la cantidad de variables temporales especificadas) numérica con los periodos que debería tener cada una de éstas.

datos

Un data frame con los microdatos.

intervalo

Vector o lista (dependiendo de la cantidad de variables temporales especificadas) numérica que contiene los períodos de inicio y fin con los cuales se quiere realizar el filtro.

textNA

Cadena de caracteres indicando el nombre que se dará a los registros cuando éstos presenten algún dato faltante (en la variable seleccionada más no en las variables temporales especificadas). El valor por defecto es "Sin Información".

ask

Si es TRUE (valor predeterminado) mostrará un mensaje por consola preguntándole al usuario cuál periodo desea que sea el último por considerar (esperando una respuesta por consola). Si previamente se ha introducido el argumento intervalo éste quedará inhabilitado y no se ejecutará.

Value

Un data frame o tibble perteneciente a las clases "tbl_df", "tbl" y "data.frame".

Examples

if (require("readxl")) {
  df1 <- read_excel(read_example("TestConsolidado1.xlsx"))
  df2 <- read_excel(read_example("TestConsolidado2.xlsx"))
}
#> Loading required package: readxl

if (FALSE) {
# Seleccione para cada dataset (A, B, C y D) una de las opciones que se le muestran
# en consola (1, 2, 3 y 4) respectivamente.
Agregar(formula = TIPO_NIVEL ~ ANO + PERIODO, frecuencia = list("Year" = 2009:2013, "Period" = 1:3), df1) -> A
Agregar(formula = TIPO_NIVEL ~ ANO + PERIODO, frecuencia = list("Year" = 2009:2013, "Period" = 1:3), df1) -> B
Agregar(formula = TIPO_NIVEL ~ ANO + PERIODO, frecuencia = list("Year" = 2009:2013, "Period" = 1:3), df1) -> C
Agregar(formula = TIPO_NIVEL ~ ANO + PERIODO, frecuencia = list("Year" = 2009:2013, "Period" = 1:3), df1) -> D
all.equal(C, D)
}

G <- Agregar(
  formula    = TIPO_NIVEL ~ ANO + PERIODO,
  frecuencia = list("Year" = 2009:2013, "Period" = 1:2),
  intervalo  = list(c(2009, 1), c(2013, 1)),
  datos      = df1
)
H <- Agregar(
  formula    = TIPO_NIVEL ~ ANO + PERIODO,
  frecuencia = list("Year" = 2009:2013, "Period" = 1:2),
  datos      = df1,
  ask        = FALSE
)
all.equal(G, H)
#> [1] TRUE

if (FALSE) {
# Observe las diferentes opciones que se le muestran por consola.
Agregar(formula = SEXO ~ YEAR + PERIODO, frecuencia = list("Year" = 2016:2021, "Period" = 1:3), df2)
}

Agregar(
  formula    = SEXO ~ YEAR + PERIODO,
  frecuencia = list("Year" = 2016:2021, "Period" = 1:3),
  intervalo  = list(c(2018, 1), c(2020, 3)),
  datos      = df2
)
#> # A tibble: 27 x 5
#>    Variable  YEAR PERIODO Clase           Total
#>    <chr>    <dbl>   <dbl> <fct>           <dbl>
#>  1 SEXO      2018       1 Hombre              1
#>  2 SEXO      2018       1 Mujer               1
#>  3 SEXO      2018       1 Sin Información     0
#>  4 SEXO      2018       2 Hombre              0
#>  5 SEXO      2018       2 Mujer               0
#>  6 SEXO      2018       2 Sin Información     0
#>  7 SEXO      2018       3 Hombre              0
#>  8 SEXO      2018       3 Mujer               0
#>  9 SEXO      2018       3 Sin Información     0
#> 10 SEXO      2019       1 Hombre              2
#> # ... with 17 more rows
Agregar(
  formula    = SEXO ~ YEAR,
  frecuencia = 2016:2021,
  intervalo  = c(2018, 2020),
  datos      = df2
)
#> # A tibble: 9 x 4
#>   Variable  YEAR Clase           Total
#>   <chr>    <dbl> <fct>           <dbl>
#> 1 SEXO      2018 Hombre              1
#> 2 SEXO      2018 Mujer               1
#> 3 SEXO      2018 Sin Información     0
#> 4 SEXO      2019 Hombre              4
#> 5 SEXO      2019 Mujer               4
#> 6 SEXO      2019 Sin Información     0
#> 7 SEXO      2020 Hombre              5
#> 8 SEXO      2020 Mujer               5
#> 9 SEXO      2020 Sin Información     0