Cree un gráfico de barras que muestra la información en columnas horizontales o verticales y para variables nominales u ordinales con dos diferentes paquetes

Esta función permite mostrar de forma interactiva un gráfico de barras verticales u horizontales cuya longitud/altura es proporcional al valor de la variable (categorías de una variable cualitativa), lo anterior para ayudar a la creación de informes descriptivos y analíticos. Dicho diagrama se puede representar usando dos diferentes librerías que son Highcharter y Plotly, las cuales usan internamente JavaScript.

Plot.Barras(
  datos,
  categoria,
  ano,
  periodo,
  freqRelativa = FALSE,
  vertical = TRUE,
  ordinal = FALSE,
  colores,
  titulo = "",
  labelEje = "Número de Graduados",
  addPeriodo = TRUE,
  textInfo = labelEje,
  libreria = c("highcharter", "plotly"),
  estilo = NULL
)

Arguments

datos

Un data frame.

categoria

Igual uso que en Plot.Series()

ano

Igual uso que en Plot.Torta()

periodo

Igual uso que en Plot.Torta()

freqRelativa

Igual uso que en Plot.Series()

vertical

Si es TRUE (valor predeterminado) indicará que la orientación del gráfico será vertical.

ordinal

Si es TRUE indicará que las categorías de la variable ingresada son ordinales (no nominales), esto con el fin de ordenar la disposición en el que se presentan en el eje del gráfico, el valor por defecto es FALSE.

colores

Igual uso que en Plot.Series()

titulo

Igual uso que en Plot.Series()

labelEje

Cadena de caracteres indicando la etiqueta del eje X o Y (dependiendo de la orientación del gráfico). Por defecto se emplea el rótulo "Número de Graduados".

addPeriodo

Igual uso que en Plot.Torta()

textInfo

Cadena de caracteres que especifica el texto que se escribe dentro de la caja de información al posar el cursor en alguna barra en el gráfico, producido por Highcharter, el valor por defecto es igual al de labelEje.

libreria

Igual uso que en Plot.Torta()

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados de acuerdo con la librería específica para graficar el plot y cuyo objetivo es personalizar pequeños detalles de ésta.

hc.Tema, hc.Credits y ply.Credits: Igual uso que en Plot.Series()

ply.Legend: Igual uso que en Plot.Torta()

Value

Retorna el diagrama de barras (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente a la clase "highchart" o "plotly".

Details

Al usar el paquete Highcharter y usar las opciones de descarga, el nombre del archivo descargado será la concatenación del plot graficado y la categoría usada, así, por ejemplo, si se graficó el diagrama de barras para la categoría "Nacionalidad" el nombre será PlotBarras_Nacionalidad.png.

Examples

Msj <- "Ac\u00e1 puede ir m\u00e1s informaci\u00f3n acerca del gr\u00e1fico."
Plot.Barras(
  datos        = ejConsolidadoGrad,
  categoria    = "NIVEL",
  ano          = 2021,
  periodo      = 1,
  freqRelativa = TRUE,
  vertical     = TRUE,
  ordinal      = TRUE,
  colores      = RColorBrewer::brewer.pal(5, "Spectral"),
  titulo       = "GRADUADOS DE ACUERDO CON EL NIVEL DE FORMACI\u00d3N",
  labelEje     = "Frecuencia Relativa<br>(% de graduados)",
  addPeriodo   = TRUE,
  textInfo     = "Porcentaje de Graduados",
  libreria     = "highcharter",
  estilo       = list(hc.Tema = 2, hc.Credits = Msj)
)

Msj <- "Año 2020, sin segregar por semestre (considerando ambos)."
Plot.Barras(
  datos     = ejConsolidadoGrad,
  categoria = "NIVEL",
  ano       = 2020,
  vertical  = FALSE,
  ordinal   = FALSE,
  colores   = RColorBrewer::brewer.pal(5, "Set2"),
  titulo    = "DISTRIBUCI\u00d3N DEL N\u00daMERO DE GRADUADOS POR NIVEL",
  labelEje  = "N\u00famero de Graduados",
  libreria  = "plotly",
  estilo    = list(
    ply.Credits = list(x = 0.45, y = 1.1, text = Msj), ply.Legend = FALSE
  )
)