Cree una serie de tiempo dinámica y flexible con tres diferentes paquetes

Esta función proporciona excelentes herramientas y opciones para la visualización de series de tiempo dinámicas con el objetivo de estudiar la evolución de una o varias variables a lo largo del tiempo. Dicha serie interactiva se puede representar usando tres diferentes librerías que son Highcharter, Plotly y Dygraph, las cuales usan internamente JavaScript.

Plot.Series(
  datos,
  categoria,
  freqRelativa = FALSE,
  invertir = FALSE,
  ylim,
  colores,
  titulo = "",
  labelX = "Periodo",
  labelY = "",
  libreria = c("highcharter", "plotly", "dygraphs"),
  estilo = NULL
)

Arguments

datos

Un data frame, no un objeto clase serie de tiempo o vector numérico.

categoria

Una variable categórica dentro del data frame ingresado en datos.

freqRelativa

Si es FALSE (valor predeterminado) la serie graficada representará las frecuencias absolutas (conteo) más no las relativas (porcentaje).

invertir

Si es FALSE (valor predeterminado) no se invertirá el eje Y. Establézcalo en TRUE si desea que en el eje Y el número más alto es el más cercano al origen.

ylim

Vector numérico que especifica el límite inferior y superior, respectivamente, del eje Y. Si no se introduce algún valor se mostrará todo el rango disponible para dicho eje.

colores

Cadena de caracteres indicando los colores con los cuales se deben colorear cada una de las series correspondiente a cada nivel del argumento categoria. Si no se introduce algún vector se usará la paleta rainbow por defecto.

titulo

Cadena de caracteres indicando el título principal del plot.

labelX

Cadena de caracteres indicando la etiqueta del eje X. Por defecto se emplea el rótulo "Periodo".

labelY

Cadena de caracteres indicando la etiqueta del eje Y.

libreria

Cadena de caracteres que indica el paquete con el cual se realizará la serie. Los valores permitidos son "highcharter" (valor predeterminado), "plotly" o "dygraphs". Los valores se emparejarán parcialmente.

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados de acuerdo con la librería especificada para graficar la serie y cuyo objetivo es personalizar pequeños detalles de ésta.

LegendTitle: Cadena de caracteres indicado un título para la leyenda (diferentes niveles del argumento categorias). Se utilizará tanto en el paquete Highcharter como en Plotly.

hc.Tema: Modifica el tema con el cual se creará la serie. Los posibles valores son un número entero entre \([1, 10]\) el cual hace referencia a diferentes temas disponibles en dicha librería (ffx, google, tufte, 538, ggplot2, economist, sandsignika, ft, superheroes y flatdark, respectivamente). El tema por defecto, al no ingresar valor alguno, es hc_theme_flat().

hc.Slider: Si es TRUE agrega un deslizador/navegador dinámico en la parte inferior de la serie. Proporciona herramientas para acercar y alejar partes de la serie, así como para desplazarse por el conjunto de datos. El valor por defecto es FALSE.

hc.BoxInfo: Si es TRUE (valor predeterminado) la información concerniente a cada punto se visualiza conjuntamente en un cuadro, o de forma individual (FALSE) al pasar el cursor.

hc.Credits: Cadena de caracteres indicando un subtítulo o etiqueta de créditos debajo del título principal.

ply.LegendPosition: Lista que especifica la posición y orientación de la leyenda. Los valores por defecto la ubican centrada verticalmente a la derecha del plot, es decir, c(x = 1, y = 0.5, orientation = "v").

ply.Interaction: Cadena de caracteres que determina el modo de las interacciones de desplazamiento. Los valores permitidos son "x unified" (valor predeterminado), "y unified", "closest", "x", "y" y FALSE.

ply.Credits: Lista que especifica la posición y texto para añadir un subtítulo o etiqueta de créditos a la serie principal, por ejemplo, c(x = 0.2, y = 1, text = "https://...").

dyg.LegendWidth: Un número que indica el ancho (en píxeles) que ocupará la leyenda. El valor por defecto es 250.

dyg.Resaltar: Si es FALSE (valor predeterminado) no se resaltará la serie en que se sitúa el cursor.

Value

Retorna la serie (objeto widget de HTML) creada. La clase del objeto retornado será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente a la clase "highchart", "plotly" o "dygraphs".

Details

Al usar el paquete Highcharter y usar las opciones de descarga, el nombre del archivo descargado será la concatenación del plot graficado y la categoría usada, así, por ejemplo, si se graficó la serie de tiempo para la categoría "Sede" el nombre será PlotSeries_Sede.png.

Examples

misColores <- c(
  "#29ABE2", # AZUL CLARO  | Amazonia
  "#8CC63F", # VERDE       | Bogota
  "#CC241D", # ROJO        | Caribe
  "#0071BC", # AZUL VIVO   | Manizales
  "#F15A24", # NARANJA     | Medellin
  "#FBB03B", # AMARILLO    | Orinoquia
  "#93278F", # MORADO      | Palmira
  "#8A381A"  # GRIS        | Tumaco
)
Msj <- "Distribuci\u00f3n de estudiantes graduados (desde el 2009-I al 2021-I) por sede."
Txt <- "EVOLUCI\u00d3N DEL N\u00daMERO DE GRADUADOS POR SEDE"
Plot.Series(
  datos        = ejConsolidadoGrad,
  categoria    = "SEDE_NOMBRE_ADM",
  freqRelativa = TRUE,
  ylim         = c(0, 75),
  colores      = misColores,
  titulo       = Txt,
  labelY       = "Frecuencia Relativa<br>(% de graduados)",
  libreria     = "highcharter",
  estilo       = list(LegendTitle = "SEDE:", hc.Tema = 10, hc.Slider = TRUE, hc.Credits = Msj)
)
#> Joining, by = c("Fecha", "Clase")

Plot.Series(
  datos     = ejConsolidadoGrad,
  categoria = "SEDE_NOMBRE_ADM",
  invertir  = TRUE,
  colores   = misColores,
  titulo    = Txt,
  labelY    = "N\u00famero de Graduados",
  libreria  = "plotly",
  estilo    = list(
    LegendTitle = "SEDE:", ply.Interaction = "closest",
    ply.LegendPosition = list(x = 0.16, y = -0.25, orientation = "h"),
    ply.Credits = list(x = 0.5, y = 0.1, text = Msj)
  )
)
#> Joining, by = c("Fecha", "Clase")

Plot.Series(
  datos     = ejConsolidadoGrad,
  categoria = "SEDE_NOMBRE_ADM",
  colores   = misColores,
  titulo    = Txt,
  labelY    = "N\u00famero de Graduados (k: miles)",
  libreria  = "dygraphs",
  estilo    = list(dyg.LegendWidth = 650, dyg.Resaltar = TRUE)
)
#> Joining, by = c("Fecha", "Clase")

# Agrupando para eliminar el semestre
if (require("dplyr")) {
  df <- ejConsolidadoGrad %>% group_by(Variable, YEAR, Clase) %>%
    summarise(Total = sum(Total, na.rm = TRUE))
}
#> `summarise()` has grouped output by 'Variable', 'YEAR'. You can override using the `.groups` argument.
Msj <- "Comportamiento anual, considerando ambos semestres (exceptuando el caso del 2021)."
Plot.Series(
  datos     = df,
  categoria = "SEXO",
  ylim      = c(1000, 6000),
  colores   = c("#3360FF", "#F30081"),
  titulo    = "EVOLUCI\u00d3N DEL N\u00daMERO DE GRADUADOS POR SEXO",
  labelX    = "A\u00f1o",
  labelY    = "N\u00famero de Graduados",
  libreria  = "highcharter",
  estilo    = list(hc.Tema = 1, hc.Credits = Msj)
)
#> Joining, by = c("Fecha", "Clase")