Cree un gráfico circular/torta/pie dinámico y flexible con dos diferentes paquetes

Esta función permite mostrar de forma interactiva una descripción compacta y general de una variable con sus respectivas categorías. Dicho diagrama se puede representar usando dos diferentes librerías que son Highcharter y Plotly, las cuales usan internamente JavaScript.

Plot.Torta(
  datos,
  categoria,
  ano,
  periodo,
  colores,
  titulo = "",
  label = "",
  addPeriodo = TRUE,
  libreria = c("highcharter", "plotly"),
  estilo = NULL
)

Arguments

datos

Un data frame.

categoria

Igual uso que en Plot.Series()

ano

Valor numérico que indica el año de interés para realizar el gráfico. Si solo se introduce el año, pero no el período, se calculará independiente del semestre (tomando ambos).

periodo

Valor numérico que indica el período o semestre de interés para realizar el gráfico. Si solo se introduce el período, pero no el año, se calculará en base a todos los años disponibles en datos.

colores

Igual uso que en Plot.Series()

titulo

Igual uso que en Plot.Series()

label

Cadena de caracteres indicando la etiqueta a la que hace referencia el plot. Por defecto se emplea el rótulo "Número de ".

addPeriodo

Si es TRUE (valor predeterminado ssi se introduce los argumentos ano y periodo simultáneamente) concatenará al título introducido una cadena de caracteres indicando entre paréntesis el "año-periodo" insertado.

libreria

Cadena de caracteres que indica el paquete con el cual se realizará el plot. Los valores permitidos son "highcharter" (valor predeterminado) y "plotly". Los valores se emparejarán parcialmente.

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados de acuerdo con la librería específica para graficar la torta y cuyo objetivo es personalizar pequeños detalles de ésta.

LegendTitle, hc.Tema, hc.Credits y ply.Credits: Igual uso que en Plot.Series()

ply.Legend: Por defecto la gráfica muestra la leyenda fuera del gráfico de pie, si se introduce la cadena de texto "inside" se resumirá toda la información dentro del pie.

Value

Retorna el diagrama circular (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente a la clase "highchart" o "plotly".

Details

Al usar el paquete Highcharter y usar las opciones de descarga, el nombre del archivo descargado será la concatenación del plot graficado y la categoría usada, así, por ejemplo, si se graficó el diagrama de pie para la categoría "Sexo" el nombre será PlotTorta_Sexo.png.

Note

Los gráficos circulares son una forma muy mala de mostrar información. El ojo es bueno para juzgar medidas lineales y malo para juzgar áreas relativas. Un gráfico de barras o un gráfico de puntos es una forma preferible de mostrar este tipo de datos.

Examples

col <- c("#F15A24", "#8CC63F")
Msj <- "Distribuci\u00f3n de estudiantes gradados en el primer periodo acad\u00e9mico del 2021."
Txt <- "DISTRIBUCI\u00d3N DE GRADUADOS POR MODALIDAD DE FORMACI\u00d3N"
Plot.Torta(
  datos     = ejConsolidadoGrad,
  categoria = "TIPO_NIVEL",
  ano       = 2021,
  periodo   = 1,
  colores   = col,
  titulo    = Txt,
  label     = "Graduados",
  libreria  = "highcharter",
  estilo    = list(
    LegendTitle = "\u00c9sta es una descripci\u00f3n para la leyenda:",
    hc.Tema = 7, hc.Credits = Msj
  )
)

Msj <- "Distribuci\u00f3n hist\u00f3rica de estudiantes gradados (desde el 2009-I al 2021-I)."
Plot.Torta(
  datos     = ejConsolidadoGrad,
  categoria = "TIPO_NIVEL",
  colores   = col,
  titulo    = Txt,
  libreria  = "plotly",
  estilo    = list(
    ply.Legend = "inside", ply.Credits = list(
      x = 0.66, y = 1.1, text = paste0("<b>", Msj, "</b>")
    )
  )
)