Cree un gráfico circular/torta/pie dinámico y flexible con dos diferentes paquetes

Esta función permite mostrar de forma interactiva una descripción compacta y general de una variable con sus respectivas categorías. Dicho diagrama se puede representar usando dos diferentes librerías que son Highcharter y Plotly, las cuales usan internamente JavaScript.

Plot.Torta(
  datos,
  categoria,
  ano,
  periodo,
  colores,
  titulo = "",
  label = "",
  addPeriodo = TRUE,
  libreria = c("highcharter", "plotly"),
  estilo = NULL
)

Argumentos

datos

Un data frame, no un objeto clase serie de tiempo o vector numérico.

categoria

Una variable categórica dentro del data frame ingresado en datos.

ano

Valor numérico que indica el año de interés para realizar el gráfico. Si solo se introduce el año, pero no el período, se calculará independiente del semestre (tomando ambos).

periodo

Valor numérico que indica el período o semestre de interés para realizar el gráfico. Si solo se introduce el período, pero no el año, se calculará con base en todos los años disponibles en datos.

colores

Cadena de caracteres indicando los colores con los cuales se deben colorear cada una de las series correspondiente a cada nivel del argumento categoria. Si no se introduce algún vector se usará la paleta rainbow por defecto.

titulo

Cadena de caracteres indicando el título principal del plot.

label

Cadena de caracteres indicando la etiqueta a la que hace referencia el plot. Por defecto se emplea el rótulo "Número de ".

addPeriodo

Si es TRUE (valor predeterminado ssi se introduce los argumentos ano y periodo simultáneamente) concatenará al título introducido una cadena de caracteres indicando entre paréntesis el "año-periodo" insertado.

libreria

Cadena de caracteres que indica el paquete con el cual se realizará el plot. Los valores permitidos son "highcharter" (valor predeterminado) y "plotly". Los valores se emparejarán parcialmente.

estilo

Lista compuesta por varios parámetros, los cuales van a ser usados de acuerdo con la librería especificada para graficar la torta y cuyo objetivo es personalizar pequeños detalles de ésta.

LegendTitle, hc.Tema, hc.Credits y ply.Credits: Igual uso que en Plot.Series()

ply.Legend: Igual uso que en Plot.Barras()

Valor

Retorna el diagrama circular (objeto widget de HTML) creado. La clase del objeto retornado será un "htmlwidget" y dependiendo de la librería usada pertenecerá adicionalmente a la clase "highchart" o "plotly".

Detalles

Al usar el paquete Highcharter y usar las opciones de descarga, el nombre del archivo descargado será la concatenación del plot graficado y la categoría usada, así, por ejemplo, si se graficó el diagrama de pie para la categoría "Sexo" el nombre será PlotTorta_Sexo.png.

Nota

Los gráficos circulares son una forma muy mala de mostrar información. El ojo es bueno para juzgar medidas lineales y malo para juzgar áreas relativas. Un gráfico de barras o un gráfico de puntos es una forma preferible de mostrar este tipo de datos.

Lista de argumentos de estilo

Sabemos que puede ser abrumador el número de argumentos dentro del parámetro estilo, pero es necesario si queremos ofrecer al usuario la máxima personalización dentro de cada función usando cualquier librería. Por tal razón, a continuación, se detalla el listado completo de argumentos, usados al especificar la librería y en qué función están presentes (marcado con una × si lo posee).

Librería	estilo$	Plot.Series()	Plot.Barras()	Plot.Apiladas()	Plot.Boxplot()	Plot.Radar()	Plot.Treemap()	Plot.Torta()	Plot.Drilldown()

┌	gg.Tema	×	×	×	×

│	gg.Texto	×	×	×	×

│	gg.Legend	×		×	×

│	gg.Linea	×

│	gg.Punto	×

│	gg.Bar		×	×

│	gg.VarWidth				×

│	gg.OutShape				×

│	gg.JitWidth				×

│	gg.JitSize				×

│	gg.Range					×

ggplot2	gg.plty					×

│	gg.plwd					×

│	gg.cglwd					×

│	gg.cglcol					×

│	gg.fontsize.title						×

│	gg.fontsize.labels						×

│	gg.fontcolor.labels						×

│	gg.border.lwds						×

│	gg.border.col						×

│	gg.lowerbound.cex.labels						×

│	gg.force.print.labels						×

└	gg.overlap.labels						×

╔	hc.Tema	×	×	×	×		×	×	×

║	hc.Credits	×	×	×	×		×	×	×

highcharter	hc.BoxInfo	×

║	hc.Slider	×

╚	hc.borderRadius						×

┌	ply.Credits	×	×	×	×	×	×	×

│	ply.Legend		×					×

│	ply.LegendPosition	×		×	×	×

plotly	ply.Interaction	×			×

│	ply.Relleno					×

│	ply.Opacidad					×	×

└	ply.LegendTitle					×

dygraphs	dyg.LegendWidth	×

╚	dyg.Resaltar	×

┌	e.Tema					×

│	e.Credits					×

echarts4r	e.Forma					×

│	e.LegType					×

└	e.LegLoc					×

Examples

col <- c("#F15A24", "#8CC63F")
Msj <- "Distribuci\u00f3n de estudiantes graduados en el primer periodo acad\u00e9mico del 2021."
Txt <- "DISTRIBUCI\u00d3N DE GRADUADOS POR MODALIDAD DE FORMACI\u00d3N"
Plot.Torta(
  datos     = ejConsolidadoGrad,
  categoria = "TIPO_NIVEL",
  ano       = 2021,
  periodo   = 1,
  colores   = col,
  titulo    = Txt,
  label     = "Graduados",
  libreria  = "highcharter",
  estilo    = list(
    LegendTitle = "\u00c9sta es una descripci\u00f3n para la leyenda:",
    hc.Tema = 7, hc.Credits = Msj
  )
)

Msj <- "Distribuci\u00f3n hist\u00f3rica de estudiantes graduados (desde el 2009-I al 2021-I)."
Plot.Torta(
  datos     = ejConsolidadoGrad,
  categoria = "TIPO_NIVEL",
  colores   = col,
  titulo    = gsub("DOS POR", "DOS\nPOR", Txt),
  libreria  = "plotly",
  estilo    = list(
    ply.Legend = "inside", ply.Credits = list(
      x = 0.8, y = 1.1, text = paste0("<b>", Msj, "</b>")
    )
  )
)