Cree fácilmente un widget para visualizar los resultados del Saber Pro en tablas HTML usando el paquete DT

Esta función está diseñada para facilitar la creación de tablas para informes y publicaciones produciendo un widget HTML para visualizar un data frame utilizando el paquete DT. La forma en que esta función maneja las cosas por usted significa que a menudo no tiene que preocuparse por los pequeños detalles para obtener un resultado impresionante y listo para usar.

Tabla.SaberPro(
  datos,
  variable,
  encabezado = "Encabezados de los Niveles de la Categoría",
  leyenda,
  tituloPdf = NULL,
  mensajePdf = "",
  ajustarNiveles = TRUE,
  colorHead = "#FFFFFF",
  colorear = FALSE,
  estilo
)

Arguments

datos
Igual uso que en Tabla()

variable
Análogo al parámetro categoria de la función Tabla()

encabezado
Igual uso que en Tabla()

leyenda
Igual uso que en Tabla() con la excepción de que, si no se introduce ningún valor, el valor por defecto será una nota explicando a que hace referencia los valores y columnas de la tabla.

tituloPdf
Igual uso que en Tabla()

mensajePdf
Igual uso que en Tabla()

ajustarNiveles
Igual uso que en Tabla()

colorHead
Igual uso que en Tabla()

colorear
Igual uso que en Tabla()

estilo
Una lista compuesta por dos parámetros:

PaletaYear: Vector de caracteres que específica los colores de fondo para los años.

PaletaCategoria: Vector de caracteres que específica los colores de fuente para las distintas categorías de la variable.

Value

Retorna la tabla creada mediante DT la cual pertenece a la clase "datatables" y "htmlwidget".

Details

Esta función se basa enteramente del paquete DT, el cual proporciona una interfaz para R a la biblioteca DataTables de JavaScript. Los data frames de R se pueden mostrar como tablas en páginas HTML, proporcionando opciones de filtrado, paginación, clasificación y muchas otras características en las tablas.

Examples

if (require("dplyr")) {
  VariosYears <- SaberPro %>% mutate(YEAR = replace(YEAR, YEAR==2019, 2020)) %>% bind_rows(SaberPro)
}
Msj <- "\u00c9sta es una descripci\u00f3n de la tabla diferente al valor por default."
Tabla.SaberPro(datos      = VariosYears,
               variable   =  "SEXO",
               encabezado = "PUNTAJES POR SEXO",
               leyenda    = Msj,
               colorHead  = "#FF5B5B",
               estilo     = list(PaletaYear = c("#F9CA00", "#F68118"),
                                 PaletaCategoria = c("#2458C5", "#F0006D", "#42C501"))
               )
Tabla.SaberPro(datos      = VariosYears,
               variable   =  "SEDE",
               encabezado = "PUNTAJES POR SEDE",
               leyenda    = Msj,
               colorHead  = "#F9CA00",
               estilo     = list(PaletaYear = c("#AEF133", "#19EE9F"),
                                 PaletaCategoria = c("#DD1C1A", "#FF6700", "#7E10DE",
                                                     "#0096F2", "#42C501"))
               )